% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxa.tree.R
\name{taxa.tree}
\alias{taxa.tree}
\title{Tree from taxa}
\usage{
taxa.tree(species)
}
\arguments{
\item{species}{a list of species that is to be included in the phylogenetic tree.}
}
\description{
Generates a tree and distance matrix from a list of species names. There are reference trees for mammals and angiosperms. The input species will be sorted by reference trees.
}
\examples{
\dontrun{
#Choose a list of string of species that you want to have in a phylogenetic tree.
taxa.tree(species = c("Canis lupus", "Canis latrans", "Acer saccharum", "Castor canadensis",
"Alces alces", "Acer rubrum",
"Vulpes vulpes", "Salix babylonica", "Odocoileus virginianus", "Betula alleghaniensis",
"Rangifer tarandus","Juniperus occidentalis"))}
}
